+++ src/game_state_machine/main_menu_state.h
@@ -9,7 +9,7 @@ namespace SAS_3D {
 		~MainMenuState();
 
 		FSMStates InitializeState(SubsystemController* subsystems, const InputState& input);
-		FSMStates UpdateState(int elapsedtime, SubsystemController* subsystems, const InputState& input);
+		FSMStates UpdateState(long long elapsedtime, SubsystemController* subsystems, const InputState& input);
 		int NextState() { return _nextstate; }
 
 	private:
